include stack trace, tweak tests a bit
